move matchers functions to matchers.ts
